moved auth functions to auth controller
